+++ .cache/74d9607a0138fd32456aee37028e4928.json
@@ -1,0 +1,1 @@
+{"id":"script.js","dependencies":[{"name":"/Users/Joshua/Desktop/Coding/encryptid.github.io/package.json","includedInParent":true,"mtime":1538243683000},{"name":"./sass/style.scss","loc":{"line":2,"column":7},"parent":"/Users/Joshua/Desktop/Coding/encryptid.github.io/script.js","resolved":"/Users/Joshua/Desktop/Coding/encryptid.github.io/sass/style.scss"}],"generated":{"js":"'use strict';\n\nrequire('./sass/style.scss');\n\nvar init = function init() {\n  return console.log('hi mom!');\n}; //fix this javascript\n\n\nwindow.addEventListener('load', init());","map":{"mappings":[{"generated":{"line":3,"column":0},"source":"script.js","original":{"line":2,"column":0}},{"generated":{"line":5,"column":0},"source":"script.js","original":{"line":4,"column":0}},{"name":"init","generated":{"line":5,"column":4},"source":"script.js","original":{"line":4,"column":6}},{"generated":{"line":5,"column":11},"source":"script.js","original":{"line":4,"column":13}},{"name":"init","generated":{"line":5,"column":20},"source":"script.js","original":{"line":4,"column":6}},{"generated":{"line":5,"column":24},"source":"script.js","original":{"line":4,"column":13}},{"generated":{"line":6,"column":0},"source":"script.js","original":{"line":4,"column":13}},{"name":"console","generated":{"line":6,"column":9},"source":"script.js","original":{"line":4,"column":19}},{"name":"log","generated":{"line":6,"column":17},"source":"script.js","original":{"line":4,"column":27}},{"generated":{"line":6,"column":20},"source":"script.js","original":{"line":4,"column":19}},{"generated":{"line":6,"column":21},"source":"script.js","original":{"line":4,"column":31}},{"generated":{"line":6,"column":30},"source":"script.js","original":{"line":4,"column":19}},{"generated":{"line":6,"column":31},"source":"script.js","original":{"line":4,"column":13}},{"generated":{"line":7,"column":0},"source":"script.js","original":{"line":4,"column":13}},{"generated":{"line":7,"column":1},"source":"script.js","original":{"line":4,"column":0}},{"generated":{"line":7,"column":3},"source":"script.js","original":{"line":1,"column":0}},{"name":"window","generated":{"line":10,"column":0},"source":"script.js","original":{"line":6,"column":0}},{"name":"addEventListener","generated":{"line":10,"column":7},"source":"script.js","original":{"line":6,"column":7}},{"generated":{"line":10,"column":23},"source":"script.js","original":{"line":6,"column":0}},{"generated":{"line":10,"column":24},"source":"script.js","original":{"line":6,"column":24}},{"generated":{"line":10,"column":30},"source":"script.js","original":{"line":6,"column":0}},{"name":"init","generated":{"line":10,"column":32},"source":"script.js","original":{"line":6,"column":32}},{"generated":{"line":10,"column":38},"source":"script.js","original":{"line":6,"column":0}}],"sources":{"script.js":"//fix this javascript\nimport './sass/style.scss';\n\nconst init = () => console.log('hi mom!');\n\nwindow.addEventListener('load', init());"},"lineCount":null}},"hash":"8691b8bf2f66ce4d33789a89ddb04433","cacheData":{"env":{}}}
