+++ .cache/969bcf87952ab443d8a22aab43d8eea1.json
@@ -1,0 +1,1 @@
+{"id":"mpVp","dependencies":[{"name":"/Users/Joshua/Desktop/Coding/encryptid.github.io/package.json","includedInParent":true,"mtime":1538243683000},{"name":"./sass/style.scss","loc":{"line":2,"column":7},"parent":"/Users/Joshua/Desktop/Coding/encryptid.github.io/script.js","resolved":"/Users/Joshua/Desktop/Coding/encryptid.github.io/sass/style.scss"}],"generated":{"js":"\"use strict\";require(\"./sass/style.scss\");var s=function(){return console.log(\"hi mom!\")};window.addEventListener(\"load\",s());","map":{"mappings":[{"source":"script.js","original":{"line":6,"column":0},"generated":{"line":1,"column":0}},{"source":"script.js","original":{"line":2,"column":0},"generated":{"line":1,"column":13}},{"source":"script.js","original":{"line":2,"column":0},"generated":{"line":1,"column":21}},{"source":"script.js","original":{"line":4,"column":0},"generated":{"line":1,"column":42}},{"source":"script.js","name":"init","original":{"line":4,"column":6},"generated":{"line":1,"column":46}},{"source":"script.js","original":{"line":4,"column":13},"generated":{"line":1,"column":48}},{"source":"script.js","name":"console","original":{"line":4,"column":19},"generated":{"line":1,"column":59}},{"source":"script.js","name":"console","original":{"line":4,"column":19},"generated":{"line":1,"column":66}},{"source":"script.js","name":"log","original":{"line":4,"column":27},"generated":{"line":1,"column":74}},{"source":"script.js","original":{"line":4,"column":31},"generated":{"line":1,"column":78}},{"source":"script.js","name":"window","original":{"line":6,"column":0},"generated":{"line":1,"column":90}},{"source":"script.js","name":"addEventListener","original":{"line":6,"column":7},"generated":{"line":1,"column":97}},{"source":"script.js","original":{"line":6,"column":24},"generated":{"line":1,"column":114}},{"source":"script.js","name":"init","original":{"line":6,"column":32},"generated":{"line":1,"column":121}}],"sources":{"script.js":"//fix this javascript\nimport './sass/style.scss';\n\nconst init = () => console.log('hi mom!');\n\nwindow.addEventListener('load', init());"},"lineCount":null}},"hash":"98ce91ee22bdd0690c12b44b1c1bc129","cacheData":{"env":{}}}
