+++ .cache/d603fbd6bf14a53b5ccde09735e119fa.json
@@ -1,0 +1,1 @@
+{"id":"sass/style.scss","dependencies":[{"name":"/Users/Joshua/Desktop/Coding/encryptid.github.io/package.json","includedInParent":true,"mtime":1538243683000},{"name":"./../images/coffeeshop_s2.jpg","dynamic":true,"parent":"/Users/Joshua/Desktop/Coding/encryptid.github.io/sass/style.scss","resolved":"/Users/Joshua/Desktop/Coding/encryptid.github.io/images/coffeeshop_s2.jpg"},{"name":"/Users/Joshua/Desktop/Coding/encryptid.github.io/sass/bio.scss","includedInParent":true,"mtime":1539115904000},{"name":"/Users/Joshua/Desktop/Coding/encryptid.github.io/sass/colors.scss","includedInParent":true,"mtime":1538324803000},{"name":"/Users/Joshua/Desktop/Coding/encryptid.github.io/sass/contact.scss","includedInParent":true,"mtime":1539105618000},{"name":"/Users/Joshua/Desktop/Coding/encryptid.github.io/sass/fonts.scss","includedInParent":true,"mtime":1538350315000},{"name":"/Users/Joshua/Desktop/Coding/encryptid.github.io/sass/footer.scss","includedInParent":true,"mtime":1539116333000},{"name":"/Users/Joshua/Desktop/Coding/encryptid.github.io/sass/header.scss","includedInParent":true,"mtime":1539127663000},{"name":"/Users/Joshua/Desktop/Coding/encryptid.github.io/sass/portfolio.scss","includedInParent":true,"mtime":1539103313000},{"name":"_css_loader","parent":"/Users/Joshua/Desktop/Coding/encryptid.github.io/sass/style.scss","resolved":"/Users/Joshua/.nvm/versions/node/v7.2.0/lib/node_modules/parcel-bundler/lib/builtins/css-loader.js"}],"generated":{"css":"@import url(\"https://fonts.googleapis.com/css?family=Montserrat:400,900|Permanent+Marker\");\nheader {\n  text-align: center;\n  font-weight: 900;\n  font-size: 3em; }\n  header .header-block {\n    height: 100vh;\n    width: 100%;\n    display: flex;\n    flex-direction: column;\n    justify-content: space-evenly;\n    background: linear-gradient(rgba(186, 112, 154, 0.45), rgba(186, 112, 154, 0.45)), url(\"ee9ab879d7d0b88da839166f618e684d.jpg\");\n    background-size: cover;\n    background-repeat: no-repeat;\n    background-color: #D0719D;\n    position: relative; }\n    header .header-block h1, header .header-block h3, header .header-block a {\n      background: none;\n      color: white;\n      font-family: \"Permanent Marker\", cursive;\n      text-decoration: none; }\n    header .header-block a {\n      animation: hover 1s ease-in-out infinite;\n      position: absolute;\n      left: 50vw;\n      bottom: 10vh; }\n  header nav {\n    max-width: 100%;\n    margin-bottom: 25px; }\n    header nav ul {\n      display: flex;\n      justify-content: space-evenly;\n      margin-top: 25px; }\n      header nav ul a {\n        list-style: none;\n        text-decoration: none; }\n      header nav ul li:hover {\n        color: white; }\n\n@keyframes hover {\n  0%, 100% {\n    bottom: 10vh; }\n  50% {\n    bottom: 12vh; } }\n\n@keyframes mhover {\n  0%, 100% {\n    bottom: 10px; }\n  50% {\n    bottom: 20px; } }\n\n@media only screen and (max-width: 600px) {\n  #arrow {\n    bottom: 10px;\n    animation: mhover 1s ease-in-out infinite; }\n  header nav ul {\n    display: block; }\n    header nav ul li {\n      margin-bottom: 1vh; } }\n\n.bio-block {\n  text-align: center;\n  border-top: 10px solid #293563;\n  width: 80vw;\n  margin: 0 auto; }\n  .bio-block h2, .bio-block h3 {\n    margin-bottom: 5vh; }\n  .bio-block h2 {\n    margin-top: 5vh; }\n  .bio-block h3 {\n    font-size: 1.2em; }\n  .bio-block .icons {\n    display: flex;\n    justify-content: space-evenly;\n    margin-bottom: 50px; }\n    .bio-block .icons .icon {\n      display: block;\n      border: 5px solid #293563;\n      border-radius: 7px;\n      background: #293563; }\n      .bio-block .icons .icon svg {\n        margin: 0 10px;\n        width: 150px;\n        fill: #D0719D;\n        background: #293563; }\n      .bio-block .icons .icon figcaption {\n        font-weight: bold;\n        margin-bottom: 10px;\n        background: #293563;\n        color: #D0719D; }\n  .bio-block p {\n    margin: 2vh auto;\n    font-size: 1.2em;\n    text-align: left; }\n\n@media only screen and (max-width: 700px) {\n  .bio-block p {\n    font-size: 2em; }\n  .bio-block .icons {\n    display: block; }\n    .bio-block .icons .icon {\n      margin: 5vh auto;\n      width: 50%;\n      min-width: 155px; }\n      .bio-block .icons .icon svg {\n        width: 135px; } }\n\n.contact-block {\n  margin: 0 auto;\n  width: 80vw;\n  border-top: 10px solid #293563;\n  margin-top: 25px;\n  display: flex;\n  flex-direction: column;\n  justify-content: space-around;\n  text-align: center;\n  max-width: 100%; }\n  .contact-block h2 {\n    margin-top: 2vh;\n    margin-bottom: 5vh; }\n  .contact-block .logos {\n    display: flex;\n    justify-content: space-evenly; }\n  .contact-block .logo {\n    width: 120px;\n    height: 120px;\n    fill: #293563; }\n  .contact-block #linkedin {\n    background: #293563;\n    fill: #D0719D;\n    border-radius: 10px; }\n\n@media only screen and (max-width: 600px) {\n  .contact-block {\n    display: block; }\n    .contact-block .logos {\n      display: block; }\n      .contact-block .logos * {\n        display: block;\n        margin: 0 auto; }\n      .contact-block .logos #octo {\n        margin-bottom: 2vh; } }\n\n.portfolio-block {\n  display: flex;\n  flex-direction: column;\n  margin: 0 auto;\n  text-align: center;\n  margin-top: 25px;\n  border-top: 10px solid #293563;\n  width: 80vw; }\n  .portfolio-block h2 {\n    margin-top: 2vh;\n    margin-bottom: 2vh; }\n  .portfolio-block figure {\n    padding-bottom: 25px;\n    margin-bottom: 25px;\n    border-bottom: 3px solid #293563; }\n    .portfolio-block figure h3 {\n      margin-bottom: 1vh; }\n    .portfolio-block figure img {\n      width: 450px;\n      max-width: 100%;\n      margin-bottom: 1vh; }\n  .portfolio-block figure:last-child {\n    border-bottom: none; }\n\nfooter {\n  height: 6vh;\n  width: 100%;\n  display: flex;\n  justify-content: space-between;\n  align-items: center;\n  background: #293563; }\n  footer * {\n    background: #293563;\n    color: #D0719D; }\n\n@media only screen and (max-width: 600px) {\n  footer {\n    display: block;\n    text-align: center;\n    height: 10vh; }\n    footer * {\n      margin: 1vh auto; }\n    footer p {\n      padding-top: 1vh; } }\n\n* {\n  margin: 0;\n  padding: 0;\n  background: #D0719D;\n  color: #293563;\n  font-family: \"Montserrat\", \"Helvetica\", sans-serif;\n  font-size: 1em;\n  scroll-behavior: smooth; }\n\nhtml, body {\n  width: 100%;\n  height: 100%; }\n\nh2 {\n  font-size: 2em; }\n","js":"\nvar reloadCSS = require('_css_loader');\nmodule.hot.dispose(reloadCSS);\nmodule.hot.accept(reloadCSS);"},"hash":"d947c7febed37f53b3fa9d9a0a5e4f25","cacheData":{"env":{}}}
