+++ .cache/fb5422b5ae42efdbda6fea9e2616a303.json
@@ -1,0 +1,1 @@
+{"id":"images/cafe.png","dependencies":[{"name":"/Users/Joshua/Desktop/Coding/encryptid.github.io/package.json","includedInParent":true,"mtime":1538243683000}],"generated":{"js":"module.exports = \"/35195ca00022356b43f4a56bbf28f748.png\";"},"hash":"b7b769f51565b5eb8151f7983a49099c","cacheData":{"env":{}}}
