+++ script.cc5eea7c.css
@@ -9,7 +9,7 @@ header {
     display: flex;
     flex-direction: column;
     justify-content: space-evenly;
-    background: linear-gradient(rgba(186, 112, 154, 0.45), rgba(186, 112, 154, 0.45)), url("coffeeshop_s2.618e684d.jpg");
+    background: linear-gradient(rgba(186, 112, 154, 0.45), rgba(186, 112, 154, 0.45)), url("coffeeshop_s2.26c8eaa4.jpg");
     background-size: cover;
     background-repeat: no-repeat;
     background-color: #D0719D;
