+++ script.css
@@ -1,1 +1,1 @@
-@import url("https://fonts.googleapis.com/css?family=Montserrat:400,900|Permanent+Marker");header{text-align:center;font-weight:900;font-size:3em}header .header-block{height:100vh;display:flex;flex-direction:column;justify-content:space-evenly;background:linear-gradient(rgba(186,112,154,.45),rgba(186,112,154,.45)),url(coffeeshop_s2.38ca5c6d.jpg);background-size:cover;background-repeat:no-repeat;background-color:#d0719d}header .header-block a,header .header-block h1,header .header-block h3{background:none;color:#fff;font-family:Permanent Marker,cursive;text-decoration:none}header .header-block a{animation:hover 1s ease-in-out infinite;position:absolute;left:50vw;bottom:10vh}header nav{max-width:100%;margin-bottom:25px}header nav ul{display:flex;justify-content:space-evenly;margin-top:25px}header nav ul a{list-style:none;text-decoration:none}header nav ul li:hover{color:#fff}@keyframes hover{0%,to{bottom:10vh}50%{bottom:12vh}}@media only screen and (max-width:600px){header nav ul{display:block}header nav ul li{margin-bottom:1vh}}.bio-block{text-align:center;border-top:10px solid #293563;width:80vw;margin:0 auto}.bio-block h2,.bio-block h3{margin-bottom:5vh}.bio-block h2{margin-top:5vh}.bio-block h3{font-size:1.2em}.bio-block .icons{display:flex;justify-content:space-evenly;margin-bottom:50px}.bio-block .icons .icon{display:block;border:5px solid #293563;border-radius:7px;background:#293563}.bio-block .icons .icon svg{margin:0 10px;width:150px;fill:#d0719d;background:#293563}.bio-block .icons .icon figcaption{font-weight:700;margin-bottom:10px;background:#293563;color:#d0719d}.bio-block p{margin:2vh auto;line-height:1.2em;text-align:left}@media only screen and (max-width:700px){.bio-block .icons{display:block}.bio-block .icons figure{margin:5vh auto;width:50%}}.contact-block{width:80vw;border-top:10px solid #293563;margin:25px auto 0;display:flex;flex-direction:column;justify-content:space-around;text-align:center;max-width:100%}.contact-block h2{margin-top:2vh;margin-bottom:5vh}.contact-block .logos{display:flex;justify-content:space-evenly}.portfolio-block{display:flex;flex-direction:column;text-align:center;margin:25px auto 0;border-top:10px solid #293563;width:80vw}.portfolio-block h2{margin-top:2vh;margin-bottom:2vh}.portfolio-block figure{padding-bottom:25px;margin-bottom:25px;border-bottom:3px solid #293563}.portfolio-block figure h3{margin-bottom:1vh}.portfolio-block figure img{width:450px;margin-bottom:1vh}.portfolio-block figure:last-child{border-bottom:none}footer{height:6vh;width:100%;display:flex;justify-content:space-between;align-items:center}footer,footer *{background:#293563}footer *{color:#d0719d}*{margin:0;padding:0;background:#d0719d;color:#293563;font-family:Montserrat,Helvetica,sans-serif;font-size:1em;scroll-behavior:smooth}body,html{width:100%;height:100%}h2{font-size:2em}
+@import url("https://fonts.googleapis.com/css?family=Montserrat:400,900|Permanent+Marker");header{text-align:center;font-weight:900;font-size:3em}header .header-block{height:100vh;width:100%;display:flex;flex-direction:column;justify-content:space-evenly;background:linear-gradient(rgba(186,112,154,.45),rgba(186,112,154,.45)),url(coffeeshop_s2.38ca5c6d.jpg);background-size:cover;background-repeat:no-repeat;background-color:#d0719d;position:relative}header .header-block a,header .header-block h1,header .header-block h3{background:none;color:#fff;font-family:Permanent Marker,cursive;text-decoration:none}header .header-block a{animation:hover 1s ease-in-out infinite;position:absolute;left:50vw;bottom:10vh}header nav{max-width:100%;margin-bottom:25px}header nav ul{display:flex;justify-content:space-evenly;margin-top:25px}header nav ul a{list-style:none;text-decoration:none}header nav ul li:hover{color:#fff}@keyframes hover{0%,to{bottom:10vh}50%{bottom:12vh}}@keyframes mhover{0%,to{bottom:10px}50%{bottom:20px}}@media only screen and (max-width:600px){#arrow{bottom:10px;animation:mhover 1s ease-in-out infinite}header nav ul{display:block}header nav ul li{margin-bottom:1vh}}.bio-block{text-align:center;border-top:10px solid #293563;width:80vw;margin:0 auto}.bio-block h2,.bio-block h3{margin-bottom:5vh}.bio-block h2{margin-top:5vh}.bio-block h3{font-size:1.2em}.bio-block .icons{display:flex;justify-content:space-evenly;margin-bottom:50px}.bio-block .icons .icon{display:block;border:5px solid #293563;border-radius:7px;background:#293563}.bio-block .icons .icon svg{margin:0 10px;width:150px;fill:#d0719d;background:#293563}.bio-block .icons .icon figcaption{font-weight:700;margin-bottom:10px;background:#293563;color:#d0719d}.bio-block p{margin:2vh auto;font-size:1.2em;text-align:left}@media only screen and (max-width:700px){.bio-block p{font-size:2em}.bio-block .icons{display:block}.bio-block .icons .icon{margin:5vh auto;width:50%;min-width:155px}.bio-block .icons .icon svg{width:135px}}.contact-block{width:80vw;border-top:10px solid #293563;margin:25px auto 0;display:flex;flex-direction:column;justify-content:space-around;text-align:center;max-width:100%}.contact-block h2{margin-top:2vh;margin-bottom:5vh}.contact-block .logos{display:flex;justify-content:space-evenly}.contact-block .logo{width:120px;height:120px;fill:#293563}.contact-block #linkedin{background:#293563;fill:#d0719d;border-radius:10px}@media only screen and (max-width:600px){.contact-block,.contact-block .logos{display:block}.contact-block .logos *{display:block;margin:0 auto}.contact-block .logos #octo{margin-bottom:2vh}}.portfolio-block{display:flex;flex-direction:column;text-align:center;margin:25px auto 0;border-top:10px solid #293563;width:80vw}.portfolio-block h2{margin-top:2vh;margin-bottom:2vh}.portfolio-block figure{padding-bottom:25px;margin-bottom:25px;border-bottom:3px solid #293563}.portfolio-block figure h3{margin-bottom:1vh}.portfolio-block figure img{width:450px;max-width:100%;margin-bottom:1vh}.portfolio-block figure:last-child{border-bottom:none}footer{height:6vh;width:100%;display:flex;justify-content:space-between;align-items:center}footer,footer *{background:#293563}footer *{color:#d0719d}@media only screen and (max-width:600px){footer{display:block;text-align:center;height:10vh}footer *{margin:1vh auto}footer p{padding-top:1vh}}*{margin:0;padding:0;background:#d0719d;color:#293563;font-family:Montserrat,Helvetica,sans-serif;font-size:1em;scroll-behavior:smooth}body,html{width:100%;height:100%}h2{font-size:2em}
